backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.55.0
tavily-python>=0.5.0
numpy>=1.26.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
